move test_query_header to tests/unit/context
