+++ core/dbt/context/manifest.py
@@ -71,13 +71,3 @@ class ManifestContext(ConfiguredContext):
     @contextproperty()
     def context_macro_stack(self):
         return self.macro_stack
-
-
-class QueryHeaderContext(ManifestContext):
-    def __init__(self, config: AdapterRequiredConfig, manifest: Manifest) -> None:
-        super().__init__(config, manifest, config.project_name)
-
-
-def generate_query_header_context(config: AdapterRequiredConfig, manifest: Manifest):
-    ctx = QueryHeaderContext(config, manifest)
-    return ctx.to_dict()
